Move Model Classes to models package
