+++ api/src/main/java/ar/edu/itba/pod/census/models/Habitant.java
@@ -1,4 +1,4 @@
-package ar.edu.itba.pod.census;
+package ar.edu.itba.pod.census.models;
 
 /**
  * Class representing the data to be processed by the system (i.e an habitant from the census).
