+++ api/src/main/java/ar/edu/itba/pod/census/serialization/HabitantSerializer.java
@@ -1,6 +1,6 @@
 package ar.edu.itba.pod.census.serialization;
 
-import ar.edu.itba.pod.census.Habitant;
+import ar.edu.itba.pod.census.models.Habitant;
 import com.hazelcast.nio.ObjectDataInput;
 import com.hazelcast.nio.ObjectDataOutput;
 import com.hazelcast.nio.serialization.StreamSerializer;
